Eliminate T-junctions.
Where do I begin?

Let's start with the tile geometry and topology. Instead of drawing a tile with a single triangle strip, it is now drawn with 5 tri strips. The biggest is in the middle and has all geometry and topolgy except the outer row/column. That set of outer rows and columns are now drawn by 4 tri-strips, one for each edge. Each edge has 2 variants: one that looks exactly like what it was before, and the other variant draws half the number of vertices on the outer edge. The first variant is used in most cases. The second variant is used to match up to neighboring tiles that have half the resolution and are coarser.

Another change is to enforce that for any pair of neighbors, they will differ by at most one level of detail. Tiles that are two or more levels of detail coarser than their neighbors will be forceably subdivided until this LOD transition constraint is satisfied. Note that when one tile is forceably subdivided, it may impact tiles beyond it as well. This forced subdivision will propagate and continue until all tiles satisfy the the LOD neighborhood constraint. In practice, this general only requires a single pass, although there are cases where more passes might be needed.
   
   The analysis of neighbors is performed by tracking which tiles meet at corners. All corners of all tiles that are used is maintained. knowing which tiles meet at a corner allows us to identify its immediate neighbors.
   
   This neighbor knowledge is then placed into the tile upon completion of tessellation for subsequent retrieval during drawing. 
   
   When rendering is performed, the tile's neighbors are examined. For every neighbor that is of lower resolution, the border mesh that touches half of the vertices will be drawn. Otherwise, the normal border mesh is drawn along that edge.