Explicitly add a second parameter to splice() calls that reflects the length of the array, rather than implicitly using the full length of the array by not specifying that parameter.
What was originally written was idiomatic Javascript. This is a common way of using splice() when the enite array is to be removed.

I refer you to Section 15.4.4.2 of the reference definition of Javascript in the document ECMA Standard ECMA-262 "ECMAScript Language Specification" so see that these are equivalent behaviors.